Add files via upload
I'm still adjusting sizedBuyers, so I have not included it in the polyMarketScans.py file. It would be helpful if you could look over it because I'm having trouble getting reasonable data out of the code. Additionally, multiBuyers.py is a prototype and will return much more data than it currently does. It's results are quite promising.